+++ frontend/app-development/queries/queries.ts
@@ -14,6 +14,7 @@ import {
   datamodelMetadataPath,
   formLayoutsPath,
   layoutSettingsPath,
+  userCurrentPath,
 } from 'app-shared/api-paths';
 import { orgsListUrl } from 'app-shared/cdn-paths';
 
@@ -32,3 +33,4 @@ export const getDatamodelsXsd = (owner, app) => get(datamodelsXsdPath(owner, app
 export const getDatamodel = (owner, app) => get(datamodelMetadataPath(owner, app));
 export const getFormLayouts = (owner, app) => get(formLayoutsPath(owner, app));
 export const getFormLayoutSettings = (owner, app) => get(layoutSettingsPath(owner, app));
+export const getUser = () => get(userCurrentPath());
