+++ frontend/dashboard/services/userService.ts
@@ -1,13 +1,6 @@
+import { User } from 'app-shared/types/User';
 import { get, post } from '../../packages/shared/src/utils/networking';
 import { userCurrentPath, userLogoutPath } from 'app-shared/api-paths';
-
-export type User = {
-  avatar_url: string;
-  email: string;
-  full_name: string;
-  id: number;
-  login: string;
-};
 
 const getCurrentUser = async (): Promise<User> => {
   return await get(`${userCurrentPath()}`);
