+++ frontend/packages/shared/src/types/User.ts
@@ -1,0 +1,7 @@
+export type User = {
+  avatar_url: string;
+  email: string;
+  full_name: string;
+  id: number;
+  login: string;
+};
